Fix for crash bug in memcpy
If toString is run on an exactly 32 character slice, memcpy will crash. It took me more hours than I care to admit to find and fix this. Hope this can be merged or at least that this PR will serve as a warning to anyone else who may run into this. :)